requirements.txt
python-telegram-bot==20.7
google-genai==1.0.0
ddgs>=9.0.0
python-dotenv==1.0.0